requirements.txt
fastapi
uvicorn
pandas
langchain
langchain-community
langgraph
chromadb
neo4j
python-dotenv
google-generativeai
pytest
pytest-benchmark
ragas
rouge-score
langchain-google-genai
google-cloud-aiplatform
google-genai
python-multipart
